问题修复

+++ ipo-service-provider/src/main/java/com/yrdce/ipo/modules/sys/entity/IpoSpoRation.java
@@ -18,9 +18,9 @@ public class IpoSpoRation {
 
 	private BigDecimal servicefee;
 
-	private Long salesAllocationratio;
+	private BigDecimal salesAllocationratio;
 
-	private Long salesRebateratio;
+	private BigDecimal salesRebateratio;
 
 	private String firmname;
 
@@ -132,19 +132,19 @@ public class IpoSpoRation {
 		this.servicefee = servicefee;
 	}
 
-	public Long getSalesAllocationratio() {
+	public BigDecimal getSalesAllocationratio() {
 		return salesAllocationratio;
 	}
 
-	public void setSalesAllocationratio(Long salesAllocationratio) {
+	public void setSalesAllocationratio(BigDecimal salesAllocationratio) {
 		this.salesAllocationratio = salesAllocationratio;
 	}
 
-	public Long getSalesRebateratio() {
+	public BigDecimal getSalesRebateratio() {
 		return salesRebateratio;
 	}
 
-	public void setSalesRebateratio(Long salesRebateratio) {
+	public void setSalesRebateratio(BigDecimal salesRebateratio) {
 		this.salesRebateratio = salesRebateratio;
 	}
